+++ R/geneticAlgorithm.R
@@ -110,13 +110,12 @@
 #'   As with omega_scale and W_scale, this argument should be adjusted carefully if specified by hand. \strong{NOTE}
 #'   that if lambdas are constrained in some other way than restricting some of them to be identical, this parameter
 #'   should be adjusted accordingly in order to the estimation succeed!
-#' @param ar_scale a positive real number adjusting how large AR parameter values are typically generated in
-#'   some random mutations: larger value implies larger coefficients (in absolute value). See the function
-#'   \code{random_coefmats2} for details. By default, a new scale is drawn from \eqn{(0,1)} uniform distribution
-#'   in each iteration. This is ignored when estimating constrained models.
-#' @param regime_force_scale a non-negative real number specifying how much should natural selection favour individuals
-#'   with less regimes that have almost all mixing weights (practically) at zero. Set to zero for no favouring or large
-#'   number for heavy favouring. Without any favouring the genetic algorithm gets more often stuck in an area of the
+#' @param ar_scale a positive real number adjusting how large AR parameter values are typically proposed in construction
+#'   of the initial population: larger value implies larger coefficients (in absolute value). After construction of the
+#'   initial population, a new scale is drawn from \eqn{(0,1)} uniform distribution in each iteration.
+#' @param regime_force_scale a non-negative real number specifying how much should natural selection favor individuals
+#'   with less regimes that have almost all mixing weights (practically) at zero. Set to zero for no favoring or large
+#'   number for heavy favoring. Without any favoring the genetic algorithm gets more often stuck in an area of the
 #'   parameter space where some regimes are wasted, but with too much favouring the best genes might never mix into
 #'   the population and the algorithm might converge poorly. Default is \code{1} and it gives \eqn{2x} larger surviving
 #'   probability weights for individuals with no wasted regimes compared to individuals with one wasted regime.
@@ -167,7 +166,7 @@
 
 GAfit <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "mean"), constraints=NULL, structural_pars=NULL,
                   ngen=200, popsize, smart_mu=min(100, ceiling(0.5*ngen)), initpop=NULL, mu_scale, mu_scale2, omega_scale, W_scale,
-                  lambda_scale, ar_scale, regime_force_scale=1, red_criteria=c(0.05, 0.01), pre_smart_mu_prob=0.05,
+                  lambda_scale, ar_scale=0.2, regime_force_scale=1, red_criteria=c(0.05, 0.01), pre_smart_mu_prob=0.05,
                   to_return=c("alt_ind", "best_ind"), minval, seed=NULL) {
 
   # Required values and premilinary checks
@@ -212,11 +211,6 @@ GAfit <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "
     W_scale <- omega_scale
   } else if(!(length(W_scale) == d & all(W_scale > 0))) {
     stop("W_scale must be numeric vector with length d and positive elements")
-  }
-  if(missing(ar_scale)) {
-    ar_scale <- 0.2 # Initial ar_scale
-  } else {
-    stopifnot(length(ar_scale == 1) && ar_scale > 0)
   }
   stopifnot(pre_smart_mu_prob >= 0 && pre_smart_mu_prob <= 1)
   if(is.null(structural_pars$C_lambda)) {
@@ -432,6 +426,7 @@ GAfit <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "
     mutate <- rbinom(n=popsize, size=1, prob=mu_rates)
     which_mutate <- which(mutate == 1)
     pre_smart_mu <- runif(1, min=1e-6, max=1-1e-6) < pre_smart_mu_prob
+    ar_scale <- runif(1, min=1e-6, max=1-1e-6) # Random AR scale
     if(i1 <= smart_mu & length(which_mutate) >= 1 & !pre_smart_mu) { # Random mutations
       if(!is.null(constraints) | runif(1, min=1e-6, max=1-1e-6) > 0.5) { # Regular, can be nonstationary
         stat_mu <- FALSE
@@ -442,7 +437,6 @@ GAfit <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "
                                                                                    structural_pars=structural_pars), numeric(npars))
       } else { # For stationarity with algorithm (slower but can skip stationarity test), Ansley and Kohn (1986)
         stat_mu <- TRUE
-        ar_scale <- runif(1, min=1e-6, max=1-1e-6) # Random AR scale
         H2[,which_mutate] <- vapply(1:length(which_mutate), function(x) random_ind2(p=p, M=M, d=d, mu_scale=mu_scale,
                                                                                     mu_scale2=mu_scale2, omega_scale=omega_scale,
                                                                                     ar_scale=ar_scale, W_scale=W_scale,