+++ NEWS.md
@@ -84,3 +84,7 @@
 * The default lags used in quantile residual tests are now 1, 3, 6, and 12.
 * Updated the documentation of fitGMVAR: more instructions on what to do when the algorithm fails to create an initial population were added.
 * Some updates in the vignette, examples, and other documentation.
+
+# gmvarkit 1.3.1 
+
+* Bug fix which required that Brobdingnag was added back to the imported packages. There was in some rare cases a problem with the exact log-likelihood function that was introduced in the version 1.3.0 (the fix was introduced in the same day as the bug, however).